refactor(loose_ends): #572 Phase 2 extract Finding to break circular import (Task 9 fixup)
