+++ src/findajob/loose_ends/walkthrough.py
@@ -10,53 +10,15 @@ LLM is only called at evaluate_dom steps to judge what's rendered.
 
 from __future__ import annotations
 
-import json
-from dataclasses import asdict, dataclass
+from dataclasses import dataclass
 from html.parser import HTMLParser
 from pathlib import Path
 from urllib.parse import urlparse
 
 import yaml
 
-
-@dataclass(frozen=True)
-class Finding:
-    """One classified loose-end candidate from a single evaluate_dom step.
-
-    Different shape from phase 1's Finding — Phase 2 carries persona,
-    walkthrough provenance, and the (persona, route, rubric) exclusion key.
-    """
-
-    persona: str  # "nux_user" | "established_user"
-    walkthrough_name: str  # matches config/loose_ends_walkthroughs.yaml
-    current_url: str  # path the walker was on when it evaluated
-    category: int  # 2 or 3
-    is_loose_end: bool  # LLM's judgment (false if excluded)
-    confidence: str  # "high" | "medium" | "low" | "review"
-    rationale: str
-    suggested_surface: str
-    excluded: bool  # true if filtered before LLM call
-    exclusion_key: str | None  # filled when excluded=True
-
-
-def write_finding(target: Path, finding: Finding) -> None:
-    """Append one JSONL row. Creates the file if needed."""
-    target.parent.mkdir(parents=True, exist_ok=True)
-    with target.open("a", encoding="utf-8") as f:
-        f.write(json.dumps(asdict(finding)) + "\n")
-
-
-def read_findings(source: Path) -> list[Finding]:
-    """Read a JSONL file into a list of Finding."""
-    if not source.exists():
-        return []
-    out: list[Finding] = []
-    for raw in source.read_text(encoding="utf-8").splitlines():
-        if not raw.strip():
-            continue
-        out.append(Finding(**json.loads(raw)))
-    return out
-
+from findajob.loose_ends.finding import Finding
+from findajob.loose_ends.rubrics import evaluate_empty_state_no_guidance, evaluate_flow_without_exit
 
 _VALID_PERSONAS = {"nux_user", "established_user"}
 
@@ -238,13 +200,6 @@ def dispatch_step(
             raise AssertionError(f"selector not present: {step.selector}")
         return None, 0.0
     if isinstance(step, EvaluateDomStep):
-        # Lazy import to break the circular dependency:
-        # rubrics.py imports Finding from this module, so top-level import would cycle.
-        from findajob.loose_ends.rubrics import (  # noqa: PLC0415
-            evaluate_empty_state_no_guidance,
-            evaluate_flow_without_exit,
-        )
-
         dom = page.content()
         current_url = _url_path(page.url)
         hints = extract_hints(dom=dom, current_url=current_url)
